+++ tools/src/librtools/ReventFd.hpp
@@ -1,6 +1,6 @@
-// $Id: ReventFd.hpp 1125 2019-03-30 07:34:54Z mueller $
+// $Id: ReventFd.hpp 1161 2019-06-08 11:52:01Z mueller $
 //
-// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-08  1161   1.1    derive from Rfd, inherit Fd
 // 2018-12-16  1084   1.0.1  use =delete for noncopyable instead of boost
 // 2013-01-14   475   1.0    Initial version
 // 2013-01-11   473   0.5    First draft
@@ -26,32 +27,24 @@
 #ifndef included_Retro_ReventFd
 #define included_Retro_ReventFd 1
 
+#include "Rfd.hpp"
+
 namespace Retro {
 
-  class ReventFd {
+  class ReventFd : public Rfd {
     public:
                     ReventFd();
-      virtual      ~ReventFd();
+      explicit      ReventFd(const char* cnam);
 
                     ReventFd(const ReventFd&) = delete;   // noncopyable 
       ReventFd&     operator=(const ReventFd&) = delete;  // noncopyable
 
-      int           Fd() const;
-      int           Signal();
-      int           Wait();
-
-      operator      int() const;
-
-      static int    SignalFd(int fd);
-      static int    WaitFd(int fd);
-
-    protected:
-
-      int           fFd;
+      void          Signal(uint64_t val=1);
+      uint64_t      Wait();
 };
   
 } // end namespace Retro
 
-#include "ReventFd.ipp"
+//#include "ReventFd.ipp"
 
 #endif