+++ tools/src/librtools/RtimerFd.cpp
@@ -1,4 +1,4 @@
-// $Id: RtimerFd.cpp 1125 2019-03-30 07:34:54Z mueller $
+// $Id: RtimerFd.cpp 1161 2019-06-08 11:52:01Z mueller $
 //
 // Copyright 2013-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-08  1161   1.1    derive from Rfd, inherit IsOpen,Close,Fd
 // 2017-02-18   852   1.0    Initial version
 // 2013-01-11   473   0.1    First draft
 // ---------------------------------------------------------------------------
@@ -43,16 +44,15 @@ namespace Retro {
 //! FIXME_docs
 
 RtimerFd::RtimerFd()
-  : fFd(-1)
+  : RtimerFd("RtimerFd::")
 {}
 
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-RtimerFd::~RtimerFd()
-{
-  Close();
-}
+RtimerFd::RtimerFd(const char* cnam)
+  : Rfd(cnam)
+{}
 
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
@@ -60,23 +60,11 @@ RtimerFd::~RtimerFd()
 void RtimerFd::Open(clockid_t clkid)
 {
   if (IsOpen())
-    throw Rexception("RtimerFd::Open()", "bad state: already open");
+    throw Rexception(fCnam+"Open()", "bad state: already open");
 
   fFd = ::timerfd_create(clkid, TFD_NONBLOCK);
   if (!IsOpen()) 
-    throw Rexception("RtimerFd::Open()", "timerfd_create() failed: ", errno);
-  return;
-}
-
-//------------------------------------------+-----------------------------------
-//! FIXME_docs
-
-void RtimerFd::Close()
-{
-  if (IsOpen()) {
-    ::close(fFd);
-    fFd = -1;
-  }
+    throw Rexception(fCnam+"Open()", "timerfd_create() failed: ", errno);
   return;
 }
 
@@ -86,11 +74,10 @@ void RtimerFd::Close()
 void RtimerFd::SetRelative(const Rtime& dt)
 {
   if (!IsOpen())
-    throw Rexception("RtimerFd::SetRelative()", "bad state: not open");
+    throw Rexception(fCnam+"SetRelative()", "bad state: not open");
 
   if (dt.Sec() <= 0 || dt.NSec() <= 0)
-    throw Rexception("RtimerFd::SetRelative()", 
-                     "bad value: dt zero or negative ");
+    throw Rexception(fCnam+"SetRelative()", "bad value: dt zero or negative ");
 
   struct itimerspec itspec;
   itspec.it_interval.tv_sec   = 0;
@@ -98,7 +85,7 @@ void RtimerFd::SetRelative(const Rtime& dt)
   itspec.it_value             = dt.Timespec();
 
   if (::timerfd_settime(fFd, 0, &itspec, nullptr) < 0)
-    throw Rexception("RtimerFd::SetRelative()", 
+    throw Rexception(fCnam+"SetRelative()", 
                      "timerfd_settime() failed: ", errno);
   return;
 }
@@ -109,7 +96,7 @@ void RtimerFd::SetRelative(const Rtime& dt)
 void RtimerFd::Cancel()
 {
   if (!IsOpen())
-    throw Rexception("RtimerFd::Cancel()", "bad state: not open");
+    throw Rexception(fCnam+"Cancel()", "bad state: not open");
 
   struct itimerspec itspec;
   itspec.it_interval.tv_sec   = 0;
@@ -119,14 +106,13 @@ void RtimerFd::Cancel()
 
   // cancel running timers
   if (::timerfd_settime(fFd, 0, &itspec, nullptr) < 0)
-    throw Rexception("RtimerFd::Cancel()", 
-                     "timerfd_settime() failed: ", errno);
+    throw Rexception(fCnam+"Cancel()", "timerfd_settime() failed: ", errno);
 
   // clear aready experied timers
   uint64_t cnt;
   int irc = ::read(fFd, &cnt, sizeof(cnt));
   if (irc < 0 && errno != EAGAIN)
-    throw Rexception("RtimerFd::Cancel()", "read() failed: ", errno);
+    throw Rexception(fCnam+"Cancel()", "read() failed: ", errno);
 
   return;
 }
@@ -137,13 +123,13 @@ void RtimerFd::Cancel()
 uint64_t RtimerFd::Read()
 {
   if (!IsOpen())
-    throw Rexception("RtimerFd::Read()", "bad state: not open");
+    throw Rexception(fCnam+"Read()", "bad state: not open");
 
   uint64_t cnt;
   int irc = ::read(fFd, &cnt, sizeof(cnt));
   if (irc < 0) {
     if (errno == EAGAIN) return 0;
-    throw Rexception("RtimerFd::Read()", "read() failed: ", errno);
+    throw Rexception(fCnam+"Read()", "read() failed: ", errno);
   }
   return cnt;
 }