+++ tools/src/librw11/Rw11CntlDEUNA.cpp
@@ -1,4 +1,4 @@
-// $Id: Rw11CntlDEUNA.cpp 1133 2019-04-19 18:43:00Z mueller $
+// $Id: Rw11CntlDEUNA.cpp 1161 2019-06-08 11:52:01Z mueller $
 //
 // Copyright 2014-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-08  1161   0.5.10 adapt to new RtimerFd API
 // 2019-04-19  1133   0.5.9  use ExecWibr()
 // 2019-02-23  1114   0.5.8  use std::bind instead of lambda
 // 2018-12-19  1090   0.5.7  use RosPrintf(bool)
@@ -251,7 +252,7 @@ Rw11CntlDEUNA::Rw11CntlDEUNA()
     fRxDscNxt{},
     fRxPollTime(0.01),
     fRxQueLimit(1000),
-    fRxPollTimer(),
+    fRxPollTimer("Rw11CntlDEUNA::fRxPollTimer."),
     fRxBufQueue(),
     fRxBufCurr(),
     fRxBufOffset(0)